tests, better transformations
transformations store rotation/translation/scaling/shearing rather than
a matrix. This is more natural and simpler. For advanced internal
calculations, the matrix form is prefered, still, hence there are
methods to convert the transformation to/from matrix.

transformations are in  translation-rotation-scaling-shearing-order
(they used to be in  translation-scaling-shearing-rotation-order).
I believe this gives better user experience.

tests to make sure the transformation to/from matrix methods do what
they are supposed to do.