represent fill path as vector of `QPainterPath`es, not one `QPainterPath`
QPainterPath can represent multiple (even disconnected) areas.
However, overlapping areas are not distinguished from non overlapping.
I.e., if the brush is semi-transparent, overlapping areas will have the same opacity as "normal" areas.
In theory it does not matter because we aim to prevent overlapping areas, however, it's useful to be able to detect them visually for monitoring and debugging.