Make scene in NodeModel optional
Change the type of the field from Scene& to Scene* and intend to allow nullptrs.
That's important for unit tests, as creating a Scene is cumbersome.
Scene is not really required in NodeModel, but it's sometimes convenient to have it (e.g., if the NodeModel is used in a NodeView), hence, having it optionally fits quite nice.