FIx PathVector copy ctor
Copying an Edge is dangerous because it has a quite some
(indirect) references to the paths it is contained in.
The copy would almost certainly be torn out of that context,
which is probably not what the user expects.

Copying Paths or Points is not a good idea for the same reason.
However, it makes sense and is required to copy PathVectors.
PathVectors are self-contained.
It is not trivial to copy a PathVector because many internal references
must be updated, but it is logically meaningful and possible.