fix PathVectorView normalization
It's not a good idea to compare the second and third edge in a
PathVectorView to decide if it needs to be reversed because
The reversed PathVectorView may still not be normalized because
after reversing other edges will be considered as second and third.
It's better to use the second and the last edge because after
reversing (while keeping the first edge fixed), second and last edge
are swapped, hence the same two edges participate in reverse-check
as before normalization.