drop PathVectorView::is_valid
There are a lot of special cases, hence the current implementation is quite
complex already.
And it misses at least one exeception:
A PathVectorView may have an arbitrary number of loops (a loop is an
edge with start=end).

I'm not sure if there are more edge cases.
The sole purpose of the PathVector::is_valid function was to be put into
some asserts to fail early.
Since we cannot rely on its correctness, it's better to not check it.

It's better to have no check than to rely on or debug with a faulty check.